fixed autodoc
Somehow, I can't make ausummary worked anymore. I removed it and generate the files with automodule "manually" for every module.
The problem either come from the relative import change in the __init__.py file and/or from the fact that the modules have the same name than one of the class inside (i.e. oimModel.py and oimModel class)