EXTVER is sometimes optional and defined to be 1 if not present
https://heasarc.gsfc.nasa.gov/docs/fcg/standard_dict.html

KEYWORD:   EXTVER
REFERENCE: FITS Standard
STATUS:    reserved
HDU:       extension
VALUE:     integer
RANGE:     [1:]
DEFAULT:   1
COMMENT:   version of the extension
DEFINITION: The value field shall contain an integer, to be used to
distinguish among different extensions in a FITS file with the same
type and name, i.e., the same values for XTENSION and EXTNAME. The
values need not start with 1 for the first extension with a particular
value of EXTNAME and need not be in sequence for subsequent values. If
the EXTVER keyword is absent, the file should be treated as if the
value were 1.  This keyword is used to describe an extension and should
not appear in the primary header.