Fix the normalization factor for 2D Gaussian
The normalization factor from Eq. (15) of Berger and Segransan (2007) is
written as sqrt(4ln2/pi)/theta, which is the normalization factor for a
1D Gaussian. In fact, we have a 2D Gaussian, and the correct
normalization factor is just the square of that, i.e. 4ln2/pi/theta^2.
Currently, in oimodeler we have written sqrt(4ln2*theta/pi), which is
wrong in many ways (theta is inside sqrt and in numerator instead of
denominator).

This commit at least gets the formula right, which apparently doesn't
matter anyway, since the whole thing seems to be normalized somewhere
else.